+++ src/rootSaga.js
@@ -1,5 +1,5 @@
 import { all } from "redux-saga/effects";
-import { watchFetchPopularMovies } from "./features/popularMoviesSaga";
+import { watchFetchPopularMovies } from "./features/movies/MovieList/popularMoviesSaga";
 
 export default function* rootSaga() {
   yield all([watchFetchPopularMovies()]);
